+++ share/objects/gold-mine-buttons.lua
@@ -1,50 +1,38 @@
--- # Gold Mine Buttons
+-- Gold Mine Buttons
+-- =================
 --
 -- The goal here is to prevent the buttons from shifting. This happens when
 -- the 'Cancel' button appears, and pushes everything to the left. It can even
 -- push something off the button row onto the middle one.
 --
--- The solution is to put permanent luck changing buttons onto the middle row,
--- thus leaving a free button slot on the button row for 'Cancel' to occupy
--- when it appears. Of course, the remaining buttons on the bottom row must be
--- shifted left.
-
-setobjecttype ('upgrades')
-
--- Increase Gem Quality:
---
--- Shift the button up a row and keep it in the first slot.
-if objectexists ('R000') then
-	modifyobject ('R000')
-
-	if currentobject () == 'R000' then
-		-- Art:
-		makechange (current, 'gbpy', 1) -- Button Position (Y)
-	end
-end
-
--- Increase Extra Chance:
---
--- Put the button in the second slot on the middle row.
-if objectexists ('R001') then
-	modifyobject ('R001')
-
-	if currentobject () == 'R001' then
-		-- Art:
-		makechange (current, 'gbpx', 1) -- Button Position (X)
-	end
-end
+-- The solution is to push all the extra chance related buttons onto the
+-- middle row. All other buttons remain in the existing positions.
 
 setobjecttype ('units')
 
 -- Shift the following buttons left one. Or, put another way, simply have
 -- them begin displaying from the left edge.
 local buttons = {
-	{ id = 'h01W' }, -- Buy Life
-	{ id = 'h03Q' }, -- Extra Chance
-	{ -- Extra Chance Slate
+	-- Buy Life:
+	{
+		id = 'h01W',
+		x = 1,
+		y = 2
+	},
+
+	-- Extra Chance:
+	{
+		id = 'h03Q',
+		x = 1,
+		y = 1
+	},
+
+	-- Extra Chance Slate:
+	{
 		id = 'h042',
-		icon = 'ReplaceableTextures\\CommandButtons\\BTNReplenishMana.blp'
+		icon = 'ReplaceableTextures\\CommandButtons\\BTNReplenishMana.blp',
+		x = 2,
+		y = 1
 	}
 }
 
@@ -54,7 +42,8 @@ for index, button in ipairs (buttons) do
 
 		if currentobject () == button.id then
 			-- Art:
-			makechange (current, 'ubpx', index - 1) -- Button Position (X)
+			makechange (current, 'ubpx', button.x) -- Button Position (X)
+			makechange (current, 'ubpy', button.y) -- Button Position (Y)
 
 			if button.icon then
 				makechange (current, 'uico', button.icon) -- Icon - Normal
@@ -91,8 +80,6 @@ if objectexists ('ACev') then
 
 		if currentobject () == button.id then
 			-- Art:
-			makechange (current, 'abpx', index - 1) -- Button Position (X)
-			makechange (current, 'abpy', 1) -- Button Position (Y)
 			makechange (current, 'aart', button.icon) -- Icon - Normal
 
 			-- Data/Stats: